refactor(chat): reorganize imports and improve file structure
- Update .eslintignore to include services and types directories
- Move errorHandler function to utils for better organization
- Adjust imports in chat files to reflect new structure
- Improve code readability by formatting action types in toast hook
