+++ src/app/api/chat/utils/error-handler.ts
@@ -1,0 +1,15 @@
+export function errorHandler(error: unknown): string {
+  if (error === null) {
+    return 'unknown error'
+  }
+
+  if (typeof error === 'string') {
+    return error
+  }
+
+  if (error instanceof Error) {
+    return error.message
+  }
+
+  return JSON.stringify(error)
+}
